Move testcomposer client
