+++ internal/http/testcomposer.go
@@ -1,4 +1,4 @@
-package testcomposer
+package http
 
 import (
 	"bytes"
@@ -27,8 +27,8 @@ var (
 	ErrJobNotFound = errors.New(msg.JobNotFound)
 )
 
-// Client service
-type Client struct {
+// TestComposer service
+type TestComposer struct {
 	HTTPClient  *http.Client
 	URL         string // e.g.) https://api.<region>.saucelabs.net
 	Credentials credentials.Credentials
@@ -68,7 +68,7 @@ type runner struct {
 }
 
 // GetSlackToken gets slack token.
-func (c *Client) GetSlackToken(ctx context.Context) (string, error) {
+func (c *TestComposer) GetSlackToken(ctx context.Context) (string, error) {
 	url := fmt.Sprintf("%s/v1/testcomposer/users/%s/settings/slack", c.URL, c.Credentials.Username)
 
 	req, err := requesth.NewWithContext(ctx, http.MethodGet, url, nil)
@@ -86,7 +86,7 @@ func (c *Client) GetSlackToken(ctx context.Context) (string, error) {
 }
 
 // StartJob creates a new job in Sauce Labs.
-func (c *Client) StartJob(ctx context.Context, opts job.StartOptions) (jobID string, isRDC bool, err error) {
+func (c *TestComposer) StartJob(ctx context.Context, opts job.StartOptions) (jobID string, isRDC bool, err error) {
 	url := fmt.Sprintf("%s/v1/testcomposer/jobs", c.URL)
 
 	opts.User = c.Credentials.Username
@@ -130,7 +130,7 @@ func (c *Client) StartJob(ctx context.Context, opts job.StartOptions) (jobID str
 	return j.JobID, j.IsRDC, nil
 }
 
-func (c *Client) doJSONResponse(req *http.Request, expectStatus int, v interface{}) error {
+func (c *TestComposer) doJSONResponse(req *http.Request, expectStatus int, v interface{}) error {
 	res, err := c.HTTPClient.Do(req)
 	if err != nil {
 		return err
@@ -146,7 +146,7 @@ func (c *Client) doJSONResponse(req *http.Request, expectStatus int, v interface
 }
 
 // Search returns metadata for the given search options opts.
-func (c *Client) Search(ctx context.Context, opts framework.SearchOptions) (framework.Metadata, error) {
+func (c *TestComposer) Search(ctx context.Context, opts framework.SearchOptions) (framework.Metadata, error) {
 	url := fmt.Sprintf("%s/v1/testcomposer/frameworks/%s", c.URL, opts.Name)
 
 	req, err := requesth.NewWithContext(ctx, http.MethodGet, url, nil)
@@ -239,7 +239,7 @@ func doRequestAsset(httpClient *http.Client, request *http.Request) error {
 }
 
 // UploadAsset uploads an asset to the specified jobID.
-func (c *Client) UploadAsset(jobID string, realDevice bool, fileName string, contentType string, content []byte) error {
+func (c *TestComposer) UploadAsset(jobID string, realDevice bool, fileName string, contentType string, content []byte) error {
 	request, err := createUploadAssetRequest(context.Background(), c.URL, c.Credentials.Username, c.Credentials.AccessKey, jobID, fileName, contentType, content)
 	if err != nil {
 		return err
@@ -248,7 +248,7 @@ func (c *Client) UploadAsset(jobID string, realDevice bool, fileName string, con
 }
 
 // Frameworks returns the list of available frameworks.
-func (c *Client) Frameworks(ctx context.Context) ([]framework.Framework, error) {
+func (c *TestComposer) Frameworks(ctx context.Context) ([]framework.Framework, error) {
 	url := fmt.Sprintf("%s/v1/testcomposer/frameworks", c.URL)
 
 	req, err := requesth.NewWithContext(ctx, http.MethodGet, url, nil)
@@ -265,7 +265,7 @@ func (c *Client) Frameworks(ctx context.Context) ([]framework.Framework, error) 
 }
 
 // Versions return the list of available versions for a specific framework and region.
-func (c *Client) Versions(ctx context.Context, frameworkName string) ([]framework.Metadata, error) {
+func (c *TestComposer) Versions(ctx context.Context, frameworkName string) ([]framework.Metadata, error) {
 	url := fmt.Sprintf("%s/v1/testcomposer/frameworks/%s/versions", c.URL, frameworkName)
 
 	req, err := requesth.NewWithContext(ctx, http.MethodGet, url, nil)
